# backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "SichrPlace API",
    "version": "1.0.0",
    "description": "API documentation for SichrPlace backend"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local server"
    }
  ],
  "paths": {
...
    "/api/booking-request": {
      "post": {
        "summary": "Create a booking request",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "apartmentId": { "type": "string" },
                  "move_in": { "type": "string", "format": "date" },
                  "move_out": { "type": "string", "format": "date" },
                  "tenant_names": { "type": "string" },
                  "reason": { "type": "string" },
                  "habits": { "type": "string" },
                  "payer": { "type": "string" },
                  "profile_link": { "type": "string" }
                },
                "required": ["apartmentId", "move_in", "move_out", "tenant_names", "reason", "payer"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Booking created" },
          "400": { "description": "Validation error" }
        }
      }
    },
    "/api/booking-requests/{apartmentId}": {
      "get": {
        "summary": "Get all booking requests for an apartment",
        "parameters": [
          {
            "name": "apartmentId",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "List of booking requests" }
        }
      }
    },
    "/api/send-message": {
      "post": {
        "summary": "Send a message",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "sender": { "type": "string" },
                  "recipient": { "type": "string" },
                  "content": { "type": "string" }
                },
                "required": ["sender", "recipient", "content"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Message sent" },
          "401": { "description": "Unauthorized" }
        }
      }
    },
...
  },
...
}